Updated User delete logic

+++ src/main/java/net/therap/leavemanagement/service/UserService.java
@@ -11,6 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.servlet.http.HttpSession;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * @author rumi.dipto
@@ -124,14 +125,27 @@ public class UserService {
         LeaveStat leaveStat = leaveStatService.findLeaveStatByUserId(userId);
         leaveStatService.delete(leaveStat);
 
+        List<Leave> leaveList = leaveService.findUserLeaveList(userId);
+        if (leaveList.size() > 0) {
+            leaveList.forEach(leave -> {
+                leaveService.delete(leave);
+            });
+        }
+
         if (user.getDesignation().equals(Designation.TEAM_LEAD)) {
-            for (UserManagement userManagement : userManagementService.findAllUserManagementByTeamLeadId(userId)) {
-                userManagementService.delete(userManagement);
+            List<UserManagement> userManagementList = userManagementService.findAllUserManagementByTeamLeadId(userId);
+            if (userManagementList.size() > 0) {
+                userManagementList.forEach(userManagement -> {
+                    userManagementService.delete(userManagement);
+                });
             }
         } else if ((user.getDesignation().equals(Designation.DEVELOPER)) ||
                 (user.getDesignation().equals(Designation.TESTER))) {
 
-            userManagementService.delete(userManagementService.findUserManagementByUserId(userId));
+            UserManagement userManagement = userManagementService.findUserManagementByUserId(userId);
+            if (Objects.nonNull(userManagement)) {
+                userManagementService.delete(userManagement);
+            }
         }
 
         userDao.delete(user);
